Added unit test and refactored
