+++ backend/app/graphql/__init__.py
@@ -3,7 +3,7 @@ import os
 
 from flask import current_app
 from .example import ExampleQueries, ExampleMutations
-from .all_users import UserQueries
+from .user_queries import UserQueries
 from .services import services
 from ..services.implementations.user_service import UserService
 from ..services.implementations.email_service import EmailService
