+++ backend/app/graphql/user_queries.py
@@ -1,5 +1,5 @@
 import graphene
-from ..graphql.services import services
+from .services import services
 from .types import QueryList, Query
 
 
@@ -17,6 +17,11 @@ class UserQueries(QueryList):
         role=graphene.String(default_value=""),
     )
 
+    user = graphene.Field(
+        User,
+        id=graphene.String(required=True),
+    )
+
     def resolve_users(self, info, first, offset, role):
         user_service = services["user_service"]
         users = user_service.get_users()
@@ -32,7 +37,7 @@ class UserQueries(QueryList):
                             role=user.role,
                         )
                     )
-            return filtered[offset : offset + first]  # noqa: E203
+            return filtered[offset: offset + first]  # noqa: E203
 
         return [
             *map(
@@ -41,6 +46,16 @@ class UserQueries(QueryList):
                     email=user.email,
                     role=user.role,
                 ),
-                users[offset : offset + first],  # noqa: E203
+                users[offset: offset + first],  # noqa: E203
             )
         ]
+
+    def resolve_user(self, info, id):
+        user_service = services["user_service"]
+        user = user_service.get_user_by_id(id)
+
+        return User(
+            name=f"{user.first_name} {user.last_name}",
+            email=user.email,
+            role=user.role,
+        )
