+++ backend/tests/graphql/test_all_users_query.py
@@ -1,4 +1,4 @@
-from app.graphql.all_users import UserQueries
+from app.graphql.user_queries import UserQueries
 import graphene
 from graphene.test import Client
 from app.resources.user_dto import UserDTO
@@ -65,6 +65,33 @@ def test_all_users_filter(mocker):
             "users": [
                 {"name": "John Doe", "email": "JohnDoe@email.com", "role": "Admin"},
             ]
+        }
+    }
+
+    assert executed == expected_result
+
+
+def test_user_by_id(mocker):
+    mock_result = UserDTO("1", "John", "Doe", "JohnDoe@email.com", "Admin")
+
+    mocker.patch("app.services.implementations.user_service.UserService.get_user_by_id", return_value=mock_result)
+
+    schema = graphene.Schema(query=UserQueries)
+    client = Client(schema)
+    executed = client.execute(
+        """{
+        user(id: "1") {
+            name
+            email
+            role
+        }}"""
+    )
+
+    expected_result = {
+        "data": {
+            "user": {
+                "name": "John Doe", "email": "JohnDoe@email.com", "role": "Admin"
+            }
         }
     }
